troca de hd

+++ exercício7.py
@@ -1,5 +1,5 @@
 aluno = input('Qual o nome do aluno? ')
-nota1 = int(input('Valor da nota 1 '))
-nota2 = int(input('Valor da nota 2 '))
+nota1 = float(input('Valor da nota 1 '))
+nota2 = float(input('Valor da nota 2 '))
 media = (nota1 + nota2)/2
 print('O aluno {}, tirou a primeira nota {} e a segunda nota {}, sua média é {}'.format(aluno, nota1, nota2, media))
