+++ exercício8.py
@@ -1,4 +1,4 @@
-metros = int(input('Quantos metros? '))
+metros = float(input('Quantos metros? '))
 centimetros = metros * 100
 milimetros = metros * 1000
 print('São {} metros, convertidos em {} centímetros e {} milímetros'.format(metros, centimetros, milimetros))
